OK SCSYNTH

+++ common/SC_AllocPool.h
@@ -188,14 +188,14 @@ private:
     void MarkBinBlock(size_t inIndex) {
         size_t word = inIndex >> 5;
         size_t bitPosition = inIndex & 31;
-        size_t bitValue = 1L << bitPosition;
+        size_t bitValue = static_cast<size_t>(1) << bitPosition;
         mBinBlocks[word] |= bitValue;
     }
 
     void ClearBinBlock(size_t inIndex) {
         size_t word = inIndex >> 5;
         size_t bitPosition = inIndex & 31;
-        size_t bitValue = 1L << bitPosition;
+        size_t bitValue = static_cast<size_t>(1) << bitPosition;
         mBinBlocks[word] &= ~bitValue;
     }
